requirements.txt
boto3==1.34.34
botocore==1.34.34
et-xmlfile==1.1.0
jmespath==1.0.1
numpy==1.26.3
openpyxl==3.1.2
pandas==2.2.0
pillow==10.2.0
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
python-dateutil==2.8.2
pytz==2024.1
s3transfer==0.10.0
six==1.16.0
tzdata==2023.4
Unidecode==1.3.8
urllib3==2.0.7
